+++ CHANGELOG.md
@@ -43,6 +43,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Fix macro creation drag handler
 - Fixed a small bug which could lead to the wrong modifiers on a running die
 - Fixed dice roll formatting in the chatlog
+- Fixed initiative display
 
 ## [v0.12.1]